moving trace handlers to cast::cmd::utils
